Add problem filtering to all problems page, support multiple queries
