+++ frontend/src/views/AllProblemsPage.tsx
@@ -4,15 +4,15 @@ import { getProblems, Problem } from '../api/Problem';
 import { LargeText } from '../components/core/Text';
 import ErrorMessage from '../components/core/Error';
 import Loading from '../components/core/Loading';
-import ProblemCard from '../components/card/ProblemCard';
 import { TextLink } from '../components/core/Link';
+import FilteredProblemList from '../components/problem/FilteredProblemList';
 
 const Content = styled.div`
   padding: 0 20%;
 `;
 
 function AllProblemsPage() {
-  const [problems, setProblems] = useState<Problem[] | null>(null);
+  const [problems, setProblems] = useState<Problem[]>([]);
   const [error, setError] = useState('');
   const [loading, setLoading] = useState(false);
 
@@ -28,15 +28,11 @@ function AllProblemsPage() {
     <Content>
       <LargeText>View All Problems</LargeText>
       <TextLink to="/game/create">Create new problem</TextLink>
+
+      <FilteredProblemList problems={problems} />
+
       { error ? <ErrorMessage message={error} /> : null }
       { loading ? <Loading /> : null }
-
-      {problems?.map((problem, index) => (
-        <ProblemCard
-          key={index}
-          problem={problem}
-        />
-      ))}
     </Content>
   );
 }
